swagger

+++ cmd/omgind/migrate/district.go
@@ -6,8 +6,8 @@ type DistrictRaw struct {
 	ID          string  `csv:"id"`
 	ParentID    *string `csv:"pid"`
 	TreeID      int32   `csv:"tree_id"`
-	TreeLeft    int32   `csv:"tree_left"`
-	TreeRight   int32   `csv:"tree_right"`
+	TreeLeft    int64   `csv:"tree_left"`
+	TreeRight   int64   `csv:"tree_right"`
 	TreeLevel   int32   `csv:"tree_level"`
 	IsLeaf      bool    `csv:"is_leaf"`
 	Name        string  `csv:"name"`
